x-wing: factor KATs into an integration test
That way we don't need to conditionally gate them on `std` or have
gating of `no_std` whatsoever.
